pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wikifuse"
version = "0.1.0"
description = "Merge Wikipedia articles across languages into one comprehensive, source-attributed page"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Gaurav Sood", email = "contact@gsood.com"},
]
keywords = ["wikipedia", "wikidata", "translation", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "click>=8.0.0",
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "wikitextparser>=0.55.0",
    "sentence-transformers>=2.2.0",
    "transformers>=4.20.0",
    "torch>=1.12.0",
    "numpy>=1.21.0",
    "beautifulsoup4>=4.11.0",
    "lxml>=4.9.0",
    "tqdm>=4.64.0",
    "python-dateutil>=2.8.0",
    "langdetect>=1.0.9",
    "openai>=1.0.0",
]

[project.scripts]
wikifuse = "wikifuse.cli:main"

[project.urls]
Homepage = "https://github.com/gojiplus/wikifuse"
Repository = "https://github.com/gojiplus/wikifuse"
Issues = "https://github.com/gojiplus/wikifuse/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["wikifuse*"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.vulture]
min_confidence = 80

[tool.pydoclint]
style = "google"
arg-type-hints-in-docstring = false
check-return-types = false
skip-checking-raises = true
check-class-attributes = false

[tool.deptry]
ignore = ["DEP002"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "vulture>=2.10",
    "deptry>=0.12.0",
    "pydoclint>=0.4.0",
    "mypy>=1.0.0",
    "types-requests>=2.28.0",
    "types-PyYAML>=6.0.0",
]
docs = [
    "sphinx>=7.0.0",
    "furo>=2024.0.0",
    "myst-parser>=2.0.0",
]
